refactor(services): move types to types folder
- Moved all type definitions from service files to a new 'types' folder.
- Improved code organization and readability.
- Easier maintenance and updates of types.
