+++ app/src/services/ModelPerformanceMonitor.ts
@@ -1,22 +1,4 @@
-export type PerfEvent = {
-  t: number;
-  label: string;
-  confidence: number;
-  requiresConfirmation?: boolean;
-  latencyMs?: number;
-  inferenceType?: 'local' | 'cloud';
-};
-
-export type PerfMetrics = {
-  windowSize: number;
-  avgConfidence: number;
-  uncertainRate: number; // fraction of events that required confirmation or were uncertain
-  avgLatencyMs: number;
-  medianLatencyMs: number;
-  framesProcessed: number;
-  framesDropped: number;
-  localVsCloudRatio: number;
-};
+import { PerfEvent, PerfMetrics } from '../types';
 
 export class ModelPerformanceMonitor {
   private baselineConfidence = 0.85; // default baseline until set explicitly
