+++ app/src/services/practiceScheduler.ts
@@ -3,14 +3,7 @@ import AsyncStorage from '@react-native-async-storage/async-storage';
 const KEY = 'practiceSchedules';
 const LAST_SHOWN_KEY = 'practiceSchedules:lastShown';
 
-export interface PracticeSchedule {
-  id: string;
-  gestureId: string;
-  hour: number; // 0-23
-  minute: number; // 0-59
-  daysOfWeek?: number[]; // 0(Sun)..6(Sat), optional for everyday
-  enabled: boolean;
-}
+import { PracticeSchedule } from '../types';
 
 function genId(): string {
   return Date.now().toString(36) + Math.random().toString(36).slice(2);
