+++ app/src/services/remoteGestureRecognitionService.ts
@@ -1,20 +1,6 @@
 import { API_URL, API_TOKEN } from '../constants';
 
-export interface GestureResult {
-  label: string;
-  confidence: number;
-}
-
-export interface RecognitionResponse {
-  result: GestureResult;
-  landmarks: Array<[number, number, number]>;
-  landmarks_px?: Array<[number, number, number]>;
-  image_size?: { width: number; height: number };
-  handedness?: string | null;
-  categories?: Array<{ name: string | null; score: number }>;
-  appLabel?: string | null;
-  appConfidence?: number | null;
-}
+import { RecognitionResponse } from '../types';
 
 export async function recognizeGestureRemotely(base64Image: string, profileId?: string): Promise<RecognitionResponse | null> {
   try {
