+++ app/src/types/AdaptivePerformanceManagerType.ts
@@ -1,0 +1,4 @@
+export type PerformanceProfile = {
+  fps: number;
+  lowPower: boolean;
+};
