+++ app/src/types/CircuitBreakerType.ts
@@ -1,0 +1,5 @@
+export enum CircuitBreakerState {
+  CLOSED,
+  OPEN,
+  HALF_OPEN,
+}
