+++ app/src/types/GestureOcclusionType.ts
@@ -1,0 +1,4 @@
+export interface OcclusionAssessment {
+  occluded: boolean;
+  hints: string[];
+}
