+++ app/src/types/LanguageManagerType.ts
@@ -1,0 +1,3 @@
+export type Language = 'en' | 'de';
+
+export type TranslationMap = Record<string, any>;
