+++ app/src/types/childSessionManagerType.ts
@@ -1,0 +1,4 @@
+export type SessionCallbacks = {
+  onEncouragement: () => void;
+  onBreak: () => void;
+};
