+++ app/src/types/crashReportingType.ts
@@ -1,0 +1,8 @@
+export interface CrashReport {
+  id: string;
+  name: string;
+  message: string;
+  stack?: string;
+  timestamp: number;
+  extra?: Record<string, unknown>;
+}
