+++ app/src/types/dataProtectionType.ts
@@ -1,0 +1,13 @@
+export interface GestureData {
+  gestureClass: string;
+  confidence: number;
+  timestamp: number;
+  sessionId: string;
+}
+
+export interface AnonymizedGestureData {
+  gestureClass: string;
+  confidence: number;
+  timestamp: number;
+  sessionId: string;
+}
