+++ app/src/types/dialogEngineType.ts
@@ -1,0 +1,4 @@
+export type LLMSuggestionResponse = {
+  nextWords: string[];
+  caregiverPhrases: string[];
+};
