+++ app/src/types/engagementTrackerType.ts
@@ -1,0 +1,10 @@
+export interface StoredStats {
+  sessions: number;
+  totalMs: number;
+}
+
+export interface EngagementStats {
+  totalSessions: number;
+  totalDurationMs: number;
+  averageDurationMs: number;
+}
