+++ app/src/types/framesType.ts
