+++ app/src/types/gdprServiceType.ts
@@ -1,0 +1,5 @@
+export interface ExportedProfileData {
+  profile: any;
+  usageStats: any[];
+  corrections: any[];
+}
