+++ app/src/types/handUtilsType.ts
@@ -1,0 +1,1 @@
+export type Triplet = [number, number, number];
