+++ app/src/types/healthScoreType.ts
@@ -1,0 +1,10 @@
+export interface HealthResult {
+  successRate: number; // 0..1
+  count: number;
+}
+
+export interface HistoricalHealthEntry {
+  date: string; // YYYY-MM-DD
+  successRate: number;
+  count: number;
+}
