+++ app/src/types/remoteHandDetectionServiceType.ts
@@ -1,0 +1,7 @@
+interface Landmark {
+  x: number;
+  y: number;
+  z: number;
+}
+
+export type Hand = Landmark[];
