+++ app/src/types/telemetryType.ts
@@ -1,0 +1,6 @@
+export interface TelemetryEvent {
+  timestamp: number;
+  latencyMs: number;
+  event?: string;
+  source?: string;
+}
