+++ app/src/types/trainingSyncType.ts
@@ -1,0 +1,3 @@
+export interface SyncProgressOptions {
+  onProgress?: (progress: number) => void;
+}
